+++ addOns/dev/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add more auth examples:
  - Login form with existing (invalid) values for the credentials.
  - A div that may obscure the login fields with all requiring scrolling.
+ - Login with a non std header and unrelated (but required) cookies.
 
 ## [0.10.0] - 2025-05-15
 ### Added